replaceUrlPlaceholders moved to AbstractApiMethod and set to private
